Virus/requirements.txt
cryptography==38.0.4
pycryptodome==3.16.0
pywin32==305
requests==2.28.1
urllib3==1.26.13
# tkinter-page==0.0.5 опционально